requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.4.1
idna==3.3
inflection==0.5.1
jsonschema==4.4.0
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
